requirements.txt
AWSIoTPythonSDK==1.5.4
boto3==1.34.56
aws-cdk-lib==2.131.0
